Attempts to delete a document now display a confirmation dialog.

+++ frontend/src/app/components/document/document.component.ts
@@ -6,11 +6,14 @@ import {FileUploadService} from '../../services/file-upload.service';
 import {DomSanitizer} from '@angular/platform-browser';
 import {Player} from '@vime/angular';
 import {saveAs} from 'file-saver';
+import {MatDialog} from '@angular/material/dialog';
+import {ConfirmDialogComponent} from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-document',
   templateUrl: './document.component.html',
-  styleUrls: ['./document.component.scss']
+  styleUrls: ['./document.component.scss'],
+  entryComponents: [ConfirmDialogComponent]
 })
 export class DocumentComponent implements OnInit, OnChanges {
 
@@ -23,7 +26,8 @@ export class DocumentComponent implements OnInit, OnChanges {
   fileObject: Blob;
   blobUrl: any;
 
-  constructor(private spaceService: SpaceService, private fileUploadService: FileUploadService, private sanitizer: DomSanitizer) {
+  constructor(private spaceService: SpaceService, private fileUploadService: FileUploadService, private sanitizer: DomSanitizer,
+              private dialog: MatDialog) {
   }
 
   @Input() space: Space;
@@ -64,26 +68,35 @@ export class DocumentComponent implements OnInit, OnChanges {
   }
 
   deleteDocument(doc: Document) {
-    console.log('del document ' + doc.id);
-    this.spaceService.deleteDocument(this.space, doc.id).subscribe(
-      () => {
-        // update frontend display if the delete is successful, no need for another GET
-        this.documentsOfSpace = this.documentsOfSpace.filter(document => document.id !== doc.id);
-        this.defaultSuccessHandling('You successfully deleted document ' + doc.name);
-
-        // if document is deleted, then delete the file related to it as well
-        this.fileUploadService.deleteFile(this.space, doc.name).subscribe(
+    const confirmDialog = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Confirm Remove Document',
+        message: 'Are you sure, you want to remove the document: ' + doc.name
+      }
+    });
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+        console.log('del document ' + doc.id);
+        this.spaceService.deleteDocument(this.space, doc.id).subscribe(
           () => {
-            console.log('File has been deleted');
+            // update frontend display if the delete is successful, no need for another GET
+            this.documentsOfSpace = this.documentsOfSpace.filter(document => document.id !== doc.id);
+            this.defaultSuccessHandling('You successfully deleted document ' + doc.name);
+
+            // if document is deleted, then delete the file related to it as well
+            this.fileUploadService.deleteFile(this.space, doc.name).subscribe(
+              () => {
+                console.log('File has been deleted');
+              },
+              error => {
+                this.defaultErrorHandling(error);
+              });
           },
           error => {
             this.defaultErrorHandling(error);
           });
-      },
-      error => {
-        this.defaultErrorHandling(error);
-      });
-
+      }
+    });
   }
   viewDocument(doc: Document) {
     console.log('view document ' + doc.id);
